+++ apps/api/src/modules/finance/finance.module.ts
@@ -2,11 +2,13 @@ import { Module } from '@nestjs/common';
 import { FinanceController } from './finance.controller';
 import { FinanceService } from './finance.service';
 import { FinanceEventHandler } from './finance.event-handler';
+import { LeaseAccountingService } from './lease-accounting.service';
+import { LeasesController } from './leases.controller';
 
 @Module({
-  controllers: [FinanceController],
-  providers: [FinanceService, FinanceEventHandler],
-  exports: [FinanceService],
+  controllers: [FinanceController, LeasesController],
+  providers: [FinanceService, FinanceEventHandler, LeaseAccountingService],
+  exports: [FinanceService, LeaseAccountingService],
 })
 export class FinanceModule {}